requirements.txt
google-adk==1.10.0
litellm==1.76.0
pandas==2.3.2
google-genai==1.31.0
chromadb==1.0.20